+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 
+## 0.4.0
+ July 15, 2022
+- Removed the refresh() function
+- Removed the class, StateXModel
+- call _states.clear() in AppStateX
+- Updated README.md
+
 ## 0.3.0
  July 11, 2022
 - Changed keyId to identifier in mixin StateListener